add CRUD operations on food items

+++ backend/controllers/foodController.js
@@ -1,7 +1,6 @@
 import { foodModel } from "../models/foodModel.js";
 import fs from "fs";
 
-
 const addFood = async (req, res) => {
   const { name, description, price, category } = req.body;
 
@@ -13,7 +12,7 @@ const addFood = async (req, res) => {
   const foodExists = await foodModel.findOne({ name });
 
   if (foodExists) {
-    fs.unlinkSync(`./uploads/${req.file.filename}`);
+   req.file &&  fs.unlinkSync(`./uploads/${req.file.filename}`);
     return res
       .status(400)
       .json({ msg: "Food with the same name already exists" });
@@ -34,40 +33,84 @@ const addFood = async (req, res) => {
 
   return res
     .status(200)
-    .json({ msg: "food item created successfully", data: foodItem });
+    .json({ msg: "food item added successfully", data: foodItem });
 };
 
-const listFood = async (req,res)=>{
-  const foodList = await foodModel.find({})
+const listFood = async (req, res) => {
+  const foodList = await foodModel.find({});
 
-  if(!foodList){
-    return res.status(400).json({msg: "Failed to fetch food list"})
+  if (!foodList) {
+    return res.status(400).json({ msg: "Failed to fetch food list" });
   }
 
-  return res.status(200).json({msg: "Food list fetched successfully", data: foodList})
-}
+  return res
+    .status(200)
+    .json({ msg: "Food list fetched successfully", data: foodList });
+};
 
-const removeFood = async (req,res)=>{
+const removeFood = async (req, res) => {
+  const foodItem = await foodModel.findById(req.params.id);
 
-  const foodItem = await foodModel.findById(req.body.id)
-
-  if(!foodItem){
-    return res.status(400).json({msg: "Food item not found"})
+  if (!foodItem) {
+    return res.status(400).json({ msg: "Food item not found" });
   }
 
-  if(foodItem.image){
-    fs.unlinkSync(`./uploads/${foodItem.image}`)
+  if (foodItem.image) {
+    fs.unlinkSync(`./uploads/${foodItem.image}`);
   }
 
+  await foodModel.findByIdAndDelete(foodItem._id);
 
-  await foodModel.findByIdAndDelete(foodItem._id)
+  return res.status(200).json({ msg: "Food item removed successfully" });
+};
 
-  return res.status(200).json({msg: "Food item removed successfully"})
+const editFood = async (req, res) => {
+  try {
+    const { id } = req.params;
 
-}
+    const newImage = req.file && req.file.filename;
 
+    const foodItem = await foodModel.findById(id);
 
+    if (!foodItem) {
+     newImage && fs.unlinkSync(`./uploads/${newImage}`);
+      return res.status(400).json({ msg: "Food item not found" });
+    }
 
+    const { name, description, category, price } = req.body;
 
+    if (!name || !description || !category || !price) {
+      newImage && fs.unlinkSync(`./uploads/${newImage}`);
+      return res.status(400).json({ msg: "All fields are required" });
+    }
 
-export { addFood ,listFood, removeFood};
+    if (foodItem.image) {
+      fs.unlinkSync(`./uploads/${foodItem.image}`);
+    }
+
+    const updatedFoodItem = await foodModel.findByIdAndUpdate(id, {
+      $set: {
+        image: newImage || "",
+        name,
+        description,
+        category,
+        price,
+      },
+    });
+
+    const newItem = await foodModel.findById(updatedFoodItem._id);
+
+    if (!newItem) {
+      return res.status(400).json({ msg: "Failed to update food item" });
+    }
+
+    return res
+      .status(200)
+      .json({ msg: "Food item updated successfully", data: newItem });
+  } catch (error) {
+    req.file && fs.unlinkSync(`uploads/${req.file.filename}`);
+    return res.status(500).json({ msg: "Server error" });
+  }
+};
+
+export { addFood, listFood, removeFood, editFood };
